+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+ - QSys Core fails more gracefully when the config path does not exist
+
 ## [11.2.0] - 2019-05-01
 ### Added
  - QSys Core Camera support